REFACTOR-#4978: align .../common/__init__.py between Dask and Ray
