+++ modin/core/execution/dask/implementations/pandas_on_dask/io/io.py
@@ -37,7 +37,7 @@ from modin.core.storage_formats.pandas.parsers import (
     PandasSQLParser,
     PandasExcelParser,
 )
-from modin.core.execution.dask.common.engine_wrapper import DaskWrapper
+from modin.core.execution.dask.common import DaskWrapper
 
 
 class PandasOnDaskIO(BaseIO):
