+++ modin/core/execution/dask/implementations/pandas_on_dask/partitioning/partition_manager.py
@@ -18,7 +18,7 @@ from dask.distributed import wait
 from modin.core.dataframe.pandas.partitioning.partition_manager import (
     PandasDataframePartitionManager,
 )
-from modin.core.execution.dask.common.engine_wrapper import DaskWrapper
+from modin.core.execution.dask.common import DaskWrapper
 from .virtual_partition import (
     PandasOnDaskDataframeColumnPartition,
     PandasOnDaskDataframeRowPartition,
